refactor(saves): extract SyncEngine class to saves/sync_engine/ package
SyncEngine owns the newest-wins matrix executor, dispatch logic,
and the I/O orchestrators that perform save transfers
(_do_download_save, _do_upload_save). The matrix-flow methods
(_sync_rom_saves, _sync_local_files, _sync_server_only_saves) and
the dispatch helpers (_dispatch_skip, _dispatch_upload,
_dispatch_sync_action, _build_sync_conflict_entry) move into
SyncEngine, along with state-mutation helpers used during sync
(_adopt_baseline_hash, _record_own_upload,
_update_file_sync_state) and the slow-path hash helper
(_get_server_save_hash, _handle_unexpected_error).

SyncEngine takes a SaveService back-reference (TYPE_CHECKING
guarded) for cross-cutting transitional callbacks
(_is_save_sync_enabled, _get_server_device_id, _get_rom_save_info,
_find_save_files, _file_md5, _resolve_retroarch_corename, _loop,
_log_debug, _rom_lock).

Public async wrappers on SaveService (pre_launch_sync,
post_exit_sync, sync_rom_saves, sync_all_saves,
resolve_sync_conflict) remain as thin orchestrators; their
internal call sites route through SyncEngine.

VersionsService and StatusService still back-ref the moved methods
via SaveService delegates that forward to SyncEngine — that
indirection will be removed in a later commit when those
sub-services are wired to take a SyncEngine reference directly.

Pure mechanical extraction; no new methods or behaviors. Issue B
(shared matrix iterator) is the next commit.

Refs #257.
